requirements.txt
numpy>=1.24.2
pandas>=2.0.0
matplotlib>=3.4
seaborn>=0.11
python-dotenv>=1.0.0
openai>=1.2.4
pyyaml>=6.0
scikit-learn>=1.2.2
click>=8.0.1
urllib3>=1.26.15
jsonschema==4.17.3
injector>=0.21.0
ijson>=3.2.3

# Code Execution related
ipykernel==6.26.0

pre-commit>=2.19.0
tenacity>=8.2.2
plotly>=5.14.1
pytest>=7.0.0
vcrpy>=5.0.0
colorama>=0.4.6

# Plugin related
langchain>=0.0.33
tabulate>=0.9.0